+++ crest/Assets/Crest/Crest/Scripts/LodData/RegisterClipSurfaceInput.cs
@@ -2,7 +2,9 @@
 
 // This file is subject to the MIT License as seen in the root of this folder structure (LICENSE)
 
+using UnityEditor;
 using UnityEngine;
+using UnityEngine.Rendering;
 
 namespace Crest
 {
@@ -12,7 +14,7 @@ namespace Crest
     /// </summary>
     [AddComponentMenu(MENU_PREFIX + "Clip Surface Input")]
     [HelpURL(Internal.Constants.HELP_URL_BASE_USER + "ocean-simulation.html" + Internal.Constants.HELP_URL_RP + "#clip-surface")]
-    public class RegisterClipSurfaceInput : RegisterLodDataInput<LodDataMgrClipSurface>
+    public partial class RegisterClipSurfaceInput : RegisterLodDataInput<LodDataMgrClipSurface>
     {
         /// <summary>
         /// The version of this asset. Can be used to migrate across versions. This value should
@@ -20,13 +22,38 @@ namespace Crest
         /// </summary>
         [SerializeField, HideInInspector]
 #pragma warning disable 414
-        int _version = 0;
+        int _version = 1;
 #pragma warning restore 414
+
+        const string k_SignedDistanceShaderPath = "Hidden/Crest/Inputs/Clip Surface/Signed Distance";
+
+        public enum Mode
+        {
+            Geometry,
+            Primitive,
+        }
+
+        // Have this match UnityEngine.PrimitiveType.
+        public enum Primitive
+        {
+            Sphere = 0,
+            Cube = 3,
+        }
 
         bool _enabled = true;
         public override bool Enabled => _enabled;
 
-        [Header("Convex Hull Options")]
+        [Header("Clip Surface Input Options")]
+
+        [Tooltip("Where the source of the clipping will come from.")]
+        [SerializeField]
+        Mode _mode = Mode.Primitive;
+
+        [Tooltip("The primitive to render (signed distance) into the simulation.")]
+        [SerializeField, Predicated("_mode", inverted: true, Mode.Primitive), DecoratedField]
+        Primitive _primitive = Primitive.Cube;
+
+        [Header("3D Clipping Options")]
 
         [Tooltip("Prevents inputs from cancelling each other out when aligned vertically. It is imperfect so custom logic might be needed for your use case.")]
         [SerializeField] bool _disableClipSurfaceWhenTooFarFromSurface = false;
@@ -51,15 +78,114 @@ namespace Crest
 
         static int sp_DisplacementSamplingIterations = Shader.PropertyToID("_DisplacementSamplingIterations");
         static int sp_ApplyToUnderwater = Shader.PropertyToID("_ApplyToUnderwater");
+        static readonly int sp_SignedDistanceShapeMatrix = Shader.PropertyToID("_SignedDistanceShapeMatrix");
 
-        private void LateUpdate()
+        Material _signedDistancedMaterial;
+        Primitive _activePrimitive;
+
+        // For rendering signed distance shapes and gizmos.
+        static Mesh s_Quad;
+        Matrix4x4 QuadMatrix
         {
-            if (OceanRenderer.Instance == null || _renderer == null)
+            get
+            {
+                var position = transform.position;
+                // Apply sea level to matrix so we can use it for rendering and gizmos.
+                position.y = OceanRenderer.Instance.SeaLevel;
+                var scale = Vector3.one * (Mathf.Max(transform.lossyScale.x, transform.lossyScale.y, transform.lossyScale.z) * 2f);
+                scale.z = 0f;
+                return Matrix4x4.TRS(position, Quaternion.Euler(90f, 0f, 0f), scale);
+            }
+        }
+
+        protected override void Start()
+        {
+            base.Start();
+
+            InitializeSignedDistanceMaterial();
+        }
+
+        protected override void Update()
+        {
+            base.Update();
+
+#if UNITY_EDITOR
+            InitializeSignedDistanceMaterial();
+#endif
+        }
+
+        void InitializeSignedDistanceMaterial()
+        {
+            if (_signedDistancedMaterial == null)
+            {
+                _signedDistancedMaterial = new Material(Shader.Find(k_SignedDistanceShaderPath));
+                _signedDistancedMaterial.hideFlags = HideFlags.HideAndDontSave;
+            }
+
+            // Could refactor using hashy.
+            if (_primitive != _activePrimitive)
+            {
+                foreach (var primitive in System.Enum.GetNames(typeof(Primitive)))
+                {
+                    _signedDistancedMaterial.DisableKeyword($"_{primitive.ToUpper()}");
+                }
+
+                _signedDistancedMaterial.EnableKeyword($"_{System.Enum.GetName(typeof(Primitive), _primitive).ToUpper()}");
+
+                _activePrimitive = _primitive;
+            }
+        }
+
+        public override void Draw(CommandBuffer buf, float weight, int isTransition, int lodIdx)
+        {
+            if (weight <= 0f)
             {
                 return;
             }
 
-            // Prevents possible conflicts since overlapping doesn't work for every case.
+            if (_mode == Mode.Primitive && _signedDistancedMaterial == null)
+            {
+                return;
+            }
+
+            if (_mode == Mode.Geometry && (_renderer == null || _material == null))
+            {
+                return;
+            }
+
+            buf.SetGlobalFloat(sp_Weight, weight);
+            buf.SetGlobalFloat(LodDataMgr.sp_LD_SliceIndex, lodIdx);
+            buf.SetGlobalVector(sp_DisplacementAtInputPosition, Vector3.zero);
+
+            if (_mode == Mode.Primitive)
+            {
+                if (s_Quad == null)
+                {
+                    s_Quad = Resources.GetBuiltinResource<Mesh>("Quad.fbx");
+                }
+
+                // Need this here or will see NullReferenceException on recompile.
+                if (_mpb == null)
+                {
+                    _mpb = new PropertyWrapperMPB();
+                }
+
+                buf.DrawMesh(s_Quad, QuadMatrix, _signedDistancedMaterial, submeshIndex: 0, shaderPass: 0, _mpb.materialPropertyBlock);
+            }
+            else
+            {
+                buf.DrawRenderer(_renderer, _material);
+            }
+        }
+
+        private void LateUpdate()
+        {
+            if (OceanRenderer.Instance == null || (_mode == Mode.Geometry && _renderer == null))
+            {
+                return;
+            }
+
+            // Prevents possible conflicts since overlapping doesn't work for every case for convex null.
             if (_disableClipSurfaceWhenTooFarFromSurface)
             {
                 var position = transform.position;
@@ -82,19 +208,30 @@ namespace Crest
 
             if (lodIdx > -1)
             {
+                // Need this here or will see NullReferenceException on recompile.
                 if (_mpb == null)
                 {
                     _mpb = new PropertyWrapperMPB();
                 }
 
-                _renderer.GetPropertyBlock(_mpb.materialPropertyBlock);
+                if (_mode == Mode.Geometry)
+                {
+                    _renderer.GetPropertyBlock(_mpb.materialPropertyBlock);
+                }
 
                 _mpb.SetInt(LodDataMgr.sp_LD_SliceIndex, lodIdx);
                 _mpb.SetInt(sp_DisplacementSamplingIterations, (int)_animatedWavesDisplacementSamplingIterations);
 
                 _mpb.SetFloat(sp_ApplyToUnderwater, _applyToUnderwater ? 1f : 0f);
 
-                _renderer.SetPropertyBlock(_mpb.materialPropertyBlock);
+                if (_mode == Mode.Geometry)
+                {
+                    _renderer.SetPropertyBlock(_mpb.materialPropertyBlock);
+                }
+                else
+                {
+                    _mpb.SetMatrix(sp_SignedDistanceShapeMatrix, Matrix4x4.TRS(transform.position, transform.rotation, transform.lossyScale).inverse);
+                }
             }
         }
 
@@ -104,9 +241,84 @@ namespace Crest
         protected override bool FeatureEnabled(OceanRenderer ocean) => ocean.CreateClipSurfaceData;
         protected override string RequiredShaderKeywordProperty => LodDataMgrClipSurface.MATERIAL_KEYWORD_PROPERTY;
         protected override string RequiredShaderKeyword => LodDataMgrClipSurface.MATERIAL_KEYWORD;
-
         protected override string MaterialFeatureDisabledError => LodDataMgrClipSurface.ERROR_MATERIAL_KEYWORD_MISSING;
         protected override string MaterialFeatureDisabledFix => LodDataMgrClipSurface.ERROR_MATERIAL_KEYWORD_MISSING_FIX;
+
+        protected override bool RendererRequired => _mode == Mode.Geometry;
+        protected override bool RendererOptional => _mode != Mode.Geometry;
+
+        // Use Unity's UV sphere mesh for gizmos as Gizmos.DrawSphere is too low resolution.
+        static Mesh s_SphereMesh;
+
+        protected new void OnDrawGizmosSelected()
+        {
+            Gizmos.color = GizmoColor;
+
+            if (_mode == Mode.Geometry)
+            {
+                if (TryGetComponent<MeshFilter>(out var mf))
+                {
+                    Gizmos.DrawWireMesh(mf.sharedMesh, 0, transform.position, transform.rotation, transform.lossyScale);
+                }
+
+                return;
+            }
+
+            if (s_Quad == null)
+            {
+                s_Quad = Resources.GetBuiltinResource<Mesh>("Quad.fbx");
+            }
+
+            // Show gizmo for quad which encompasses the shape.
+            Gizmos.matrix = QuadMatrix;
+            Gizmos.DrawWireMesh(s_Quad);
+
+            Gizmos.matrix = transform.localToWorldMatrix;
+
+            switch (_primitive)
+            {
+                case Primitive.Sphere:
+                    if (s_SphereMesh == null)
+                    {
+                        s_SphereMesh = Resources.GetBuiltinResource<Mesh>("New-Sphere.fbx");
+                    }
+
+                    // Render mesh and wire sphere at default size (0.5m radius) which is scaled by gizmo matrix.
+                    Gizmos.DrawMesh(s_SphereMesh, submeshIndex: 0, Vector3.zero, Quaternion.identity, Vector3.one);
+                    Gizmos.DrawWireSphere(Vector3.zero, 0.5f);
+                    break;
+                case Primitive.Cube:
+                    // Render mesh and wire box at default size which is scaled by gizmo matrix.
+                    Gizmos.DrawCube(Vector3.zero, Vector3.one);
+                    Gizmos.DrawWireCube(Vector3.zero, Vector3.one);
+                    break;
+                default:
+                    Debug.LogError("Crest: Not a valid primitive type!");
+                    break;
+            }
+        }
 #endif
+    }
+
+    // Version handling - perform data migration after data loaded.
+    public partial class RegisterClipSurfaceInput : ISerializationCallbackReceiver
+    {
+        public void OnBeforeSerialize()
+        {
+            // Intentionally left empty.
+        }
+
+        public void OnAfterDeserialize()
+        {
+            // Version 1 (2021.07.25)
+            // - default mode changed from geo to primitive
+            if (_version == 0)
+            {
+                // The user is using geometry for clipping.
+                _mode = Mode.Geometry;
+
+                _version = 1;
+            }
+        }
     }
 }