+++ crest/Assets/Crest/Crest/Scripts/LodData/RegisterHeightInput.cs
@@ -4,6 +4,7 @@
 
 using UnityEditor;
 using UnityEngine;
+using UnityEngine.Experimental.Rendering;
 
 namespace Crest
 {
@@ -12,7 +13,7 @@ namespace Crest
     /// </summary>
     [ExecuteAlways]
     [AddComponentMenu(MENU_PREFIX + "Height Input")]
-    public class RegisterHeightInput : RegisterLodDataInputWithSplineSupport<LodDataMgrAnimWaves>
+    public partial class RegisterHeightInput : RegisterLodDataInputWithSplineSupport<LodDataMgrAnimWaves>
     {
         /// <summary>
         /// The version of this asset. Can be used to migrate across versions. This value should
@@ -76,4 +77,30 @@ namespace Crest
         protected override bool FeatureEnabled(OceanRenderer ocean) => true;
 #endif // UNITY_EDITOR
     }
+
+#if UNITY_EDITOR
+    public partial class RegisterHeightInput
+    {
+        public override bool Validate(OceanRenderer ocean, ValidatedHelper.ShowMessage showMessage)
+        {
+            var isValid = base.Validate(ocean, showMessage);
+
+            if (isValid)
+            {
+                if (ocean != null && ocean._simSettingsAnimatedWaves._renderTextureGraphicsFormat != GraphicsFormat.R32G32B32A32_SFloat)
+                {
+                    showMessage(
+                        "Changing the height of the ocean can reduce precision leading to artefacts like tearing or incorrect normals. " +
+                        $"{ocean._simSettingsAnimatedWaves._renderTextureGraphicsFormat} may not have enough precision.",
+                        "Change graphics format to <i>R32G32B32A32_SFloat</i>.",
+                        ValidatedHelper.MessageType.Warning,
+                        ocean
+                    );
+                }
+            }
+
+            return isValid;
+        }
+    }
+#endif
 }