+++ crest/Assets/Crest/Crest/Scripts/Shapes/FFT/FFTCompute.cs
@@ -4,6 +4,7 @@
 
 // Inspired by https://github.com/speps/GX-EncinoWaves
 
+using System.Collections.Generic;
 using UnityEditor;
 using UnityEngine;
 using UnityEngine.Experimental.Rendering;
@@ -47,13 +48,28 @@ namespace Crest
         ComputeShader _shaderSpectrum;
         ComputeShader _shaderFFT;
 
-        float _prevWindTurbulence;
-        float _prevWindSpeed;
-        float _prevWindDir;
-        int _prevResolution;
-
         int _kernelSpectrumInit;
         int _kernelSpectrumUpdate;
+
+        // Generation data
+        int _resolution;
+        float _windSpeed;
+        float _windTurbulence;
+        float _windDirRad;
+        OceanWaveSpectrum _spectrum;
+
+        float _generationTime = -1f;
+
+        public FFTCompute(int resolution, float windSpeed, float windTurbulence, float windDirRad, OceanWaveSpectrum spectrum)
+        {
+            Debug.Assert(Mathf.NextPowerOfTwo(resolution) == resolution, "FFTCompute resolution must be power of 2");
+
+            _resolution = resolution;
+            _windSpeed = windSpeed;
+            _windTurbulence = windTurbulence;
+            _windDirRad = windDirRad;
+            _spectrum = spectrum;
+        }
 
         public void Release()
         {
@@ -82,16 +98,11 @@ namespace Crest
                 _waveBuffers = null;
             }
 
-            _prevWindTurbulence = -1;
-            _prevWindSpeed = -1;
-            _prevResolution = -1;
             _isInitialised = false;
         }
 
-        void InitializeTextures(int resolution, OceanWaveSpectrum spectrum)
+        void InitializeTextures()
         {
-            Debug.Assert(Mathf.NextPowerOfTwo(resolution) == resolution, "FFTCompute resolution must be power of 2");
-
             Release();
 
             _shaderSpectrum = Resources.Load<ComputeShader>("FFT/FFTSpectrum");
@@ -99,12 +110,12 @@ namespace Crest
             _kernelSpectrumUpdate = _shaderSpectrum.FindKernel("SpectrumUpdate");
             _shaderFFT = Resources.Load<ComputeShader>("FFT/FFTCompute");
 
-            _texButterfly = new Texture2D(resolution, Mathf.RoundToInt(Mathf.Log(resolution, 2)), TextureFormat.RGBAFloat, false, true);
+            _texButterfly = new Texture2D(_resolution, Mathf.RoundToInt(Mathf.Log(_resolution, 2)), TextureFormat.RGBAFloat, false, true);
 
             _texSpectrumControls = new Texture2D(Crest.OceanWaveSpectrum.NUM_OCTAVES, 1, TextureFormat.RFloat, false, true);
 
             RenderTextureDescriptor rtd = new RenderTextureDescriptor();
-            rtd.width = rtd.height = resolution;
+            rtd.width = rtd.height = _resolution;
             rtd.dimension = TextureDimension.Tex2DArray;
             rtd.enableRandomWrite = true;
             rtd.depthBufferBits = 0;
@@ -132,7 +143,7 @@ namespace Crest
             _tempFFT3.Create();
 
             // Raw wave data buffer
-            _waveBuffers = new RenderTexture(resolution, resolution, 0, GraphicsFormat.R16G16B16A16_SFloat);
+            _waveBuffers = new RenderTexture(_resolution, _resolution, 0, GraphicsFormat.R16G16B16A16_SFloat);
             _waveBuffers.wrapMode = TextureWrapMode.Repeat;
             _waveBuffers.antiAliasing = 1;
             _waveBuffers.filterMode = FilterMode.Bilinear;
@@ -144,49 +155,132 @@ namespace Crest
             _waveBuffers.enableRandomWrite = true;
             _waveBuffers.Create();
 
-            InitializeButterfly(resolution);
+            InitializeButterfly(_resolution);
 
-            InitialiseSpectrumHandControls(spectrum);
+            InitialiseSpectrumHandControls();
 
             _isInitialised = true;
+        }
+
+        static Dictionary<int, FFTCompute> _generators = new Dictionary<int, FFTCompute>();
+
+        static int CalculateWaveConditionsHash(int resolution, float windTurbulence, float windDirRad, float windSpeed, OceanWaveSpectrum spectrum)
+        {
+            var conditionsHash = Hashy.CreateHash();
+            Hashy.AddInt(resolution, ref conditionsHash);
+            Hashy.AddFloat(windSpeed, ref conditionsHash);
+            Hashy.AddFloat(windTurbulence, ref conditionsHash);
+            Hashy.AddFloat(windDirRad, ref conditionsHash);
+            if (spectrum != null)
+            {
+                Hashy.AddObject(spectrum, ref conditionsHash);
+            }
+            return conditionsHash;
         }
 
         /// <summary>
         /// Computes water surface displacement, with wave components split across slices of the output texture array
         /// </summary>
-        public RenderTexture GenerateDisplacements(CommandBuffer buf, int resolution, float windTurbulence, float windSpeed, float windDirRad, float time, OceanWaveSpectrum spectrum, bool updateSpectrum)
+        public static RenderTexture GenerateDisplacements(CommandBuffer buf, int resolution, float windTurbulence, float windDirRad, float windSpeed, float time, OceanWaveSpectrum spectrum, bool updateSpectrum)
         {
-            if (!_isInitialised || _spectrumHeight == null || _waveBuffers.width != resolution)
+            // All static data arguments should be hashed here and passed to the generator constructor
+            var conditionsHash = CalculateWaveConditionsHash(resolution, windTurbulence, windDirRad, windSpeed, spectrum);
+            FFTCompute generator;
+            if (!_generators.TryGetValue(conditionsHash, out generator))
             {
-                InitializeTextures(resolution, spectrum);
+                // No generator for these params - create one
+                generator = new FFTCompute(resolution, windSpeed, windTurbulence, windDirRad, spectrum);
+                _generators.Add(conditionsHash, generator);
+            }
+
+            // The remaining dynamic data arguments should be passed in to the generation here
+            return generator.GenerateDisplacementsInternal(buf, time, updateSpectrum);
+        }
+
+        RenderTexture GenerateDisplacementsInternal(CommandBuffer buf, float time, bool updateSpectrum)
+        {
+            // Check if already generated, and we're not being asked to re-update the spectrum
+            if (_generationTime == time)
+            {
+                return _waveBuffers;
+            }
+
+            if (!_isInitialised || _spectrumHeight == null)
+            {
+                InitializeTextures();
             }
 
             if (!_spectrumInitialised || updateSpectrum)
             {
-                InitialiseSpectrumHandControls(spectrum);
+                InitialiseSpectrumHandControls();
+                InitializeSpectrum(buf);
+                _spectrumInitialised = true;
             }
 
-            if (!Mathf.Approximately(_prevWindTurbulence, windTurbulence) ||
-                !Mathf.Approximately(_prevWindSpeed, windSpeed) ||
-                !Mathf.Approximately(_prevWindDir, windDirRad) ||
-                resolution != _prevResolution ||
-                updateSpectrum
-                )
-            {
-                _prevWindTurbulence = windTurbulence;
-                _prevWindSpeed = windSpeed;
-                _prevWindDir = windDirRad;
-                _prevResolution = resolution;
+            UpdateSpectrum(buf, time);
 
-                InitializeSpectrum(buf, resolution, windSpeed, windDirRad, windTurbulence, spectrum._gravityScale * Mathf.Abs(Physics.gravity.magnitude));
-            }
+            DispatchFFT(buf);
 
-            UpdateSpectrum(buf, resolution, time, spectrum._chop);
-
-            DispatchFFT(buf, resolution);
+            _generationTime = time;
 
             return _waveBuffers;
         }
+
+        /// <summary>
+        /// Changing wave gen data can result in creating lots of new generators. This gives a way to notify
+        /// that a parameter has changed. If there is no existing generator for the new param values, but there
+        /// is one for the old param values, this old generator is repurposed.
+        /// </summary>
+        public static void OnGenerationDataUpdated(int resolution,
+            float windTurbulenceOld, float windDirRadOld, float windSpeedOld, OceanWaveSpectrum spectrumOld,
+            float windTurbulenceNew, float windDirRadNew, float windSpeedNew, OceanWaveSpectrum spectrumNew)
+        {
+            // If multiple wave components share one FFT, then one of them changes its settings, it will
+            // actually steal the generator from the rest. Then the first from the rest which request the
+            // old settings will trigger creation of a new generator, and the remaining ones will use this
+            // new generator. In the end one new generator is created, but it's created for the old settings.
+            // Generators are requested single threaded so there should not be a race condition. Odd pattern
+            // but I don't think any other way works without ugly checks to see if old generators are still
+            // used, or other complicated things.
+
+            // Check if no generator exists for new values
+            var newHash = CalculateWaveConditionsHash(resolution, windTurbulenceNew, windDirRadNew, windSpeedNew, spectrumNew);
+            if (!_generators.TryGetValue(newHash, out _))
+            {
+                // Try to adapt an existing generator rather than default to creating a new one
+                var oldHash = CalculateWaveConditionsHash(resolution, windTurbulenceOld, windDirRadOld, windSpeedOld, spectrumOld);
+                FFTCompute generator;
+                if (_generators.TryGetValue(oldHash, out generator))
+                {
+                    // Hash will change for this generator, so remove the current one
+                    _generators.Remove(oldHash);
+
+                    // Update params
+                    generator._windTurbulence = windTurbulenceNew;
+                    generator._windDirRad = windDirRadNew;
+                    generator._windSpeed = windSpeedNew;
+                    generator._spectrum = spectrumNew;
+
+                    // Trigger generator to re-init the spectrum
+                    generator._spectrumInitialised = false;
+
+                    // Re-add with new hash
+                    _generators.Add(newHash, generator);
+                }
+            }
+            else
+            {
+                // There is already a new generator which will be used. Remove the previous one - if it really is needed
+                // then it will be created later.
+                var oldHash = CalculateWaveConditionsHash(resolution, windTurbulenceOld, windDirRadOld, windSpeedOld, spectrumOld);
+                _generators.Remove(oldHash);
+            }
+        }
+
+        /// <summary>
+        /// Number of FFT generators
+        /// </summary>
+        public static int GeneratorCount => _generators != null ? _generators.Count : 0;
 
         /// <summary>
         /// Computes the offsets used for the FFT calculation
@@ -227,56 +321,54 @@ namespace Crest
             _texButterfly.Apply();
         }
 
-        void InitialiseSpectrumHandControls(OceanWaveSpectrum spectrum)
+        void InitialiseSpectrumHandControls()
         {
             for (var i = 0; i < OceanWaveSpectrum.NUM_OCTAVES; i++)
             {
-                float pow = spectrum._powerDisabled[i] ? 0f : Mathf.Pow(10f, spectrum._powerLog[i]);
-                pow *= spectrum._multiplier * spectrum._multiplier;
+                float pow = _spectrum._powerDisabled[i] ? 0f : Mathf.Pow(10f, _spectrum._powerLog[i]);
+                pow *= _spectrum._multiplier * _spectrum._multiplier;
                 _spectrumDataScratch[i] = pow * Color.white;
             }
 
             _texSpectrumControls.SetPixels(_spectrumDataScratch);
             _texSpectrumControls.Apply();
-
-            _spectrumInitialised = true;
         }
 
         /// <summary>
         /// Computes base spectrum values based on wind speed & turbulence & spectrum controls
         /// </summary>
-        void InitializeSpectrum(CommandBuffer buf, int size, float windSpeed, float windDirRad, float windTurbulence, float gravity)
+        void InitializeSpectrum(CommandBuffer buf)
         {
-            buf.SetComputeIntParam(_shaderSpectrum, "_Size", size);
-            buf.SetComputeFloatParam(_shaderSpectrum, "_WindSpeed", windSpeed);
-            buf.SetComputeFloatParam(_shaderSpectrum, "_Turbulence", windTurbulence);
-            buf.SetComputeFloatParam(_shaderSpectrum, "_Gravity", gravity);
-            buf.SetComputeVectorParam(_shaderSpectrum, "_WindDir", new Vector2(Mathf.Cos(windDirRad), Mathf.Sin(windDirRad)));
+            buf.SetComputeIntParam(_shaderSpectrum, "_Size", _resolution);
+            buf.SetComputeFloatParam(_shaderSpectrum, "_WindSpeed", _windSpeed);
+            buf.SetComputeFloatParam(_shaderSpectrum, "_Turbulence", _windTurbulence);
+            buf.SetComputeFloatParam(_shaderSpectrum, "_Gravity", _spectrum._gravityScale * Mathf.Abs(Physics.gravity.magnitude));
+            buf.SetComputeVectorParam(_shaderSpectrum, "_WindDir", new Vector2(Mathf.Cos(_windDirRad), Mathf.Sin(_windDirRad)));
             buf.SetComputeTextureParam(_shaderSpectrum, _kernelSpectrumInit, "_SpectrumControls", _texSpectrumControls);
             buf.SetComputeTextureParam(_shaderSpectrum, _kernelSpectrumInit, "_ResultInit", _spectrumInit);
-            buf.DispatchCompute(_shaderSpectrum, _kernelSpectrumInit, size / 8, size / 8, CASCADE_COUNT);
+            buf.DispatchCompute(_shaderSpectrum, _kernelSpectrumInit, _resolution / 8, _resolution / 8, CASCADE_COUNT);
         }
 
         /// <summary>
         /// Computes a spectrum for the current time which can be FFT'd into the final surface
         /// </summary>
-        void UpdateSpectrum(CommandBuffer buf, int size, float time, float chop)
+        void UpdateSpectrum(CommandBuffer buf, float time)
         {
             buf.SetComputeFloatParam(_shaderSpectrum, "_Time", time);
-            buf.SetComputeFloatParam(_shaderSpectrum, "_Chop", chop);
+            buf.SetComputeFloatParam(_shaderSpectrum, "_Chop", _spectrum._chop);
             buf.SetComputeTextureParam(_shaderSpectrum, _kernelSpectrumUpdate, "_Init0", _spectrumInit);
             buf.SetComputeTextureParam(_shaderSpectrum, _kernelSpectrumUpdate, "_ResultHeight", _spectrumHeight);
             buf.SetComputeTextureParam(_shaderSpectrum, _kernelSpectrumUpdate, "_ResultDisplaceX", _spectrumDisplaceX);
             buf.SetComputeTextureParam(_shaderSpectrum, _kernelSpectrumUpdate, "_ResultDisplaceZ", _spectrumDisplaceZ);
-            buf.DispatchCompute(_shaderSpectrum, _kernelSpectrumUpdate, size / 8, size / 8, CASCADE_COUNT);
+            buf.DispatchCompute(_shaderSpectrum, _kernelSpectrumUpdate, _resolution / 8, _resolution / 8, CASCADE_COUNT);
         }
 
         /// <summary>
         /// FFT the spectrum into surface displacements
         /// </summary>
-        void DispatchFFT(CommandBuffer buf, int resolution)
+        void DispatchFFT(CommandBuffer buf)
         {
-            var kernelOffset = 2 * Mathf.RoundToInt(Mathf.Log(resolution / FFT_KERNEL_0_RESOLUTION, 2f));
+            var kernelOffset = 2 * Mathf.RoundToInt(Mathf.Log(_resolution / FFT_KERNEL_0_RESOLUTION, 2f));
 
             buf.SetComputeTextureParam(_shaderFFT, kernelOffset, "_InputH", _spectrumHeight);
             buf.SetComputeTextureParam(_shaderFFT, kernelOffset, "_InputX", _spectrumDisplaceX);
@@ -285,17 +377,23 @@ namespace Crest
             buf.SetComputeTextureParam(_shaderFFT, kernelOffset, "_Output1", _tempFFT1);
             buf.SetComputeTextureParam(_shaderFFT, kernelOffset, "_Output2", _tempFFT2);
             buf.SetComputeTextureParam(_shaderFFT, kernelOffset, "_Output3", _tempFFT3);
-            buf.DispatchCompute(_shaderFFT, kernelOffset, 1, resolution, CASCADE_COUNT);
+            buf.DispatchCompute(_shaderFFT, kernelOffset, 1, _resolution, CASCADE_COUNT);
 
             buf.SetComputeTextureParam(_shaderFFT, kernelOffset + 1, "_InputH", _tempFFT1);
             buf.SetComputeTextureParam(_shaderFFT, kernelOffset + 1, "_InputX", _tempFFT2);
             buf.SetComputeTextureParam(_shaderFFT, kernelOffset + 1, "_InputZ", _tempFFT3);
             buf.SetComputeTextureParam(_shaderFFT, kernelOffset + 1, "_InputButterfly", _texButterfly);
             buf.SetComputeTextureParam(_shaderFFT, kernelOffset + 1, "_Output", _waveBuffers);
-            buf.DispatchCompute(_shaderFFT, kernelOffset + 1, resolution, 1, CASCADE_COUNT);
+            buf.DispatchCompute(_shaderFFT, kernelOffset + 1, _resolution, 1, CASCADE_COUNT);
         }
 
-        internal void OnGUI()
+        public static void OnGUI(int resolution, float windTurbulence, float windDirRad, float windSpeed, OceanWaveSpectrum spectrum)
+        {
+            _generators.TryGetValue(CalculateWaveConditionsHash(resolution, windTurbulence, windDirRad, windSpeed, spectrum), out var generator);
+            generator?.OnGUIInternal();
+        }
+
+        void OnGUIInternal()
         {
             OceanDebugGUI.DrawTextureArray(_waveBuffers, 8, 0.5f, 20f);